[add] add a module csr
