+++ src/cpu.rs
@@ -1,6 +1,7 @@
 pub mod fetch;
 pub mod decode;
 pub mod execution;
+mod csr;
 mod instruction;
 
 use crate::bus;
@@ -47,28 +48,6 @@ impl CPU {
         let dist = dist.unwrap();
         if dist != 0 {
             self.regs[dist] = src;
-        }
-    }
-
-    pub fn read_csr(&self, src: Option<usize>) -> u32 {
-        self.csrs[src.unwrap()]
-    }
-
-    pub fn write_csr(&mut self, dist: Option<usize>, src: i32) {
-        self.csrs[dist.unwrap()] = src as u32;
-    }
-
-    pub fn bitset_csr(&mut self, dist: Option<usize>, src: i32) {
-        let mask = src as u32;
-        if mask != 0 {
-            self.csrs[dist.unwrap()] |= mask;
-        }
-    }
-
-    pub fn bitclr_csr(&mut self, dist: Option<usize>, src: i32) {
-        let mask = src as u32;
-        if mask != 0 {
-            self.csrs[dist.unwrap()] &= !mask;
         }
     }
 }
