+++ src/cpu/csr.rs
@@ -1,0 +1,25 @@
+use super::CPU;
+
+impl CPU {
+    pub fn read_csr(&self, src: Option<usize>) -> u32 {
+        self.csrs[src.unwrap()]
+    }
+
+    pub fn write_csr(&mut self, dist: Option<usize>, src: i32) {
+        self.csrs[dist.unwrap()] = src as u32;
+    }
+
+    pub fn bitset_csr(&mut self, dist: Option<usize>, src: i32) {
+        let mask = src as u32;
+        if mask != 0 {
+            self.csrs[dist.unwrap()] |= mask;
+        }
+    }
+
+    pub fn bitclr_csr(&mut self, dist: Option<usize>, src: i32) {
+        let mask = src as u32;
+        if mask != 0 {
+            self.csrs[dist.unwrap()] &= !mask;
+        }
+    }
+}
